check if states exists first

+++ 0x0F-python-object_relational_mapping/13-model_state_delete_a.py
@@ -24,8 +24,9 @@ def statesprint(username, password, database):
     states = session.query(State).\
         filter(State.name.like('%a%')).\
         order_by(State.id.asc()).all()
-    for state in states:
-        session.delete(state)
+    if states:
+        for state in states:
+            session.delete(state)
     session.commit()
     session.close
